frem_sim is main frem simulation, partgen has main partition generators
The basic idea behind frem_sim is as follows: pass in a DictRep Hamiltonian,
annealing parameters, and partition generator. The simulation then performs
diagonlization, forward, and reverse annealing to start. Finally, it performs
frem annealing on every state passed in via the partition geneator.

The output consists of a raw data (csv) file and a bulk data analysis (txt)
file.