major software reoganization for clarity
As my software has progressed, it's identity as "dwaveutils" was wavering.
I have changed the name to qanic (Quantum Annealing Nic) to relfect that it
is the culmination of my quantum annealing related research software.

To reflect this change, (and to organize the clutter), I reoganized the
placement of many functions by creating a new package structure.