massive changes to organization and some functionality changes
Since the numerical portion of this project has grown, it has been harder
to maintain and make changes due to the original organizaion. Because of this,
I have made some  major changes and changed the name of the package to
qanic (quantum annealing) nic to reflect these changes. Here are a list
of major changes.

1.) ProbRep/DictRep --> IsingH
* I changed the logic of "problem representation" to something more concrete,
an "Ising Hamiltonian." Further, this class is a lot simpler. I realized
that the parameter sweeping utility was very clunky and unnecessary. If I
have a set of parameters to sweep over, I just just do that user-side.
* In particular, annealing with numerics is now only a single run at a time.
* Further, numeric_anneal now does both reverse/forward annealing.
* User is no longer required to make anneal schedule literally...
* etc...

2.) Code is now broken into two main portions: IsingH and user-side.
* IsingH contains the code to implement an Ising Hamiltonian. Utils contains
a number of utilities necessary to do useful things, but these are not
intended to be used by a user--they are purely internal.
* On the other hand, qanic/graphing and qanic/numerics contains code that
is intended to be used by a user. For example, the numerics has a hamgen.py
file that allows a user to quickly generate a Hamiltonian.

3.) frem_sim changes
* Frem_Sim now takes more fine-granined annealing parameters.
* frem_sim now takes as argument the type of measurement scheme being used

TODOs:
1.) Add more measurement schemes.
2.) Allow numerical forward anneal to be longer than reverse anneal for frem.
3.) Create more interesting metrics for frem simulation than gs probability.
4.) Perhaps make file naming easier to deal with.