simplified make_numeric_schedule and made more robust at the same time
The old function was overly complicated for silly reasons, and as such
was a pain to debugg. But much worse, it was very rigid. If I wanted to
completely rethink how an anneal should be performed (as is happening with
frem tests), it was impossible to do easily with the old way.

The new methodology is simple. A user inputs an anneals schedule in the D-Wave
fasion of the form [[t0, s0], [t1, s1], ..., [tf, 1]] and a discretization.
A "numeric schedule" is created of the form [[list of times], [list of svals]]
where the times are spaced in accordance with discretization. The disc
of svals is then fixed by the slope relationship between t and s implicit in
making the annealing schedule.