minor name changes from measurements to revstate
Basically, there are ways to initialize the reverse annealing state for
reverse of FREM annealing that does not require a measurement, i.e. one could
instead use direct diagonlaization and then encode the actual superposition
state.

Because of this, I didn't want to have utilties written in a thing called
'measurements' that were not measurements, but I didn't want to make two
separate utility files for the 'same operation.'

Thus, I settled on calling it 'revstate' short for reverse state
initialization.