fixed wrong Hamiltonian FREM bug and changed init state to pass qutip obj
The IsingH class initializes an Hx and Hz Hamiltonian term.
However, since my HR and HF partitions include all qubits to avoid any issues
with 'skipping qubits' and the eventual final states, the Hx initilization
was including a sigma_x term on every qubit in both the R and F partition.
I have no idea what this does to the physics of FREM--but it doesn't matter;
I was doing an anneal on the WRONG Hamiltonian w.r.t. to the ground-states I
began with. Because of this, all results prior to this change are wrong.

In an effort to add more support for superposition states as things progress,
I also changed the way in which initial states are passed around. Rather than
being strings/ ints that specify the state that IsingH uses to create the
qutip objects, IsingH now expects that a reverse/FREM anneal is passed a dict
that actually contains the qutip state objects.

However, this only works if the qubits in HR and HF are partitioned
contiguosuly, so I will need to change this further to make it work for doing
arbitrary superposition states in FREM.