changed frem_sim to less general sidon_frem_sim with better file I/O
This change is a huge quality of life update for running the simulation
on HPC hardware. In effect, it prints the output of the entire parameter-sweeped
simulation to a single hdf5 file now. This can then be corralled into a
pandas data frame afterwards easily for graphing and whatnot.

Further, this simulation is made to be slightly less general than the previous
frem_sim to reflect the frequency of this type of call that I make to the MSU
supercomputer. In effect, I have less annoying python set-up that has be
done to submit a frem simulation.