removed dwave-ocean-sdk dependency for now
There is a conflict between QuTiP and dwave-ocean-sdk packages. Since the
D-Wave portion of the code is not being used and is not fleshed out, I have
decided to drop it for now.

Obviously, I hope to add support for D-Wave again in the future, so for now,
I have not removed the lines of code that support it--simply commented them.