added timing, built-in 95_5 to sidon_frem_sim and made compat with hamgen
Timing can now be toggled on/off when running a simulation by altering the
profile tag. This outputs a h5dump row of data that contains the length in
seconds it took to run forward/reverse/frem annealing with a given set of
parameters. Further, the default number of runs for a Hamiltonian of size n
is now Hsize[n] where Hsize is a dict that maps n to samples needed for a
95% confidence on a 5% interval assuming Guassian statistics.