tested Hx, -Hx debacle with forward anneal; added sidon0Kn to mix
In discussing with Novotny, we discovered there was a discrepency between
the ground-state of H(0) between my manuscript and the D-Wave docs. It turns
out, I defined the transverse field term with a positive sign whereas D-Wave
uses an overall minus sign. Hence, they initilize in the |+> state whereas I
was inializing in the |-> state.

As expected (at least for forward annealing), the two strategies yield the
same ground-state probabilities in a closed-system setting. It remains to be
seen if the same is true with reverse annealing and FREM which I will test
next.

In addition, I wanted to go ahead and make a small change to sidon_frem_sim as
a reminder that any future runs really should be over the sidon0Kn set instead
of the sidonKn set. This is because including 0 as a possible coupling means
that we get to implicitely sample over graphs which are not fully connected.
Given my re-understanding of FREM (guess on small subset where you know
solution in isolation, i.e. planted solution), I'd expect that highly connected
(but not fully) would give the best advantage for FREM over regular annealing.