updated imports, added hx_ver to sidon_frem_sim, fixed run_sidon
The old import structure required a lot of "import sys" and using relative
paths. This is obviously clunky, so I used what I learned in developing
the edd package to update the imports and use pip3 to do a local install.
This meant also adding a requirements and setup.py to the repo as well.

As for hx_ver to sidon--this is just the method by which I can test whether
the -Hx or Hx makes a difference for FREM annealing as well.

Finally, for the fixed_run_sidon, I just mean that I added the new parameters
to the file and made the filename more flexible.