upated sidon_frem_sim/ run_sidon to have more informative filenames
The filenames have two changes in particular.
1. The raw data file names are prepended with the summary data files name
and then appended with information about particular run.
2. The summary file names (and consequently the raw data file name) includes
the date of when simulation was started. This makes it easier to run sims
of the same parameters and save data to same directory without worrying about
overwriting previous data.

In addition, I updated the estimate-resources a bit with information obtained
from running the simulation on the Discovery HPC at USC.