changed reverse annealing protocol initialization schemes
In particular, I added support for the following schemes
1. Choosing the ground-state by direct diagonlization
2. Guessing ground-state by a coin toss
3. Sampling ground-state from a probability mass function

As for 3, it's generally assumed that such a pmf is obtained from a forward
anneal but this need not be the case in general.

Having added these options to the revstate utility file, I changed the
sidon_frem_sim to try all three when performing a reverse anneal. The
output file has this change reflected as well, of course.